Refactor tangible assets to follow the proposed design pattern
