+++ src/main/java/uk/gov/companieshouse/api/accounts/enumeration/AccountsResource.java
@@ -6,7 +6,8 @@ import uk.gov.companieshouse.api.accounts.links.SmallFullLinkType;
 
 public enum AccountsResource {
 
-    SMALL_FULL_STOCKS(AccountType.SMALL_FULL, Resource.STOCKS, Parent.SMALL_FULL, SmallFullLinkType.STOCKS_NOTE, "small-full-accounts#stocks-note");
+    SMALL_FULL_STOCKS(AccountType.SMALL_FULL, Resource.STOCKS, Parent.SMALL_FULL, SmallFullLinkType.STOCKS_NOTE, "small-full-accounts#stocks-note"),
+    SMALL_FULL_TANGIBLE_ASSETS(AccountType.SMALL_FULL, Resource.TANGIBLE_ASSETS, Parent.SMALL_FULL, SmallFullLinkType.TANGIBLE_ASSETS_NOTE, "small-full-accounts-note#tangible-assets");
 
     private AccountType accountType;
     private Resource resource;
