+++ src/main/java/uk/gov/companieshouse/api/accounts/model/rest/smallfull/notes/tangible/Depreciation.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.api.accounts.model.rest.notes.tangible;
+package uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.tangible;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
