+++ src/main/java/uk/gov/companieshouse/api/accounts/model/rest/smallfull/notes/tangible/TangibleAssets.java
@@ -1,15 +1,15 @@
-package uk.gov.companieshouse.api.accounts.model.rest.notes.tangible;
+package uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.tangible;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import javax.validation.Valid;
 import javax.validation.constraints.Size;
-import uk.gov.companieshouse.api.accounts.model.rest.RestObject;
+import uk.gov.companieshouse.api.accounts.model.rest.Rest;
 import uk.gov.companieshouse.api.accounts.validation.CharSetValid;
 import uk.gov.companieshouse.charset.CharSet;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class TangibleAssets extends RestObject {
+public class TangibleAssets extends Rest {
 
     private static final int MAX_FIELD_LENGTH = 20000;
     private static final int MIN_FIELD_LENGTH = 1;
