+++ src/main/java/uk/gov/companieshouse/api/accounts/validation/smallfull/TangibleAssetsValidator.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.api.accounts.validation;
+package uk.gov.companieshouse.api.accounts.validation.smallfull;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,25 +9,28 @@ import javax.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import uk.gov.companieshouse.api.accounts.enumeration.AccountsResource;
 import uk.gov.companieshouse.api.accounts.exception.DataException;
 import uk.gov.companieshouse.api.accounts.exception.ServiceException;
 import uk.gov.companieshouse.api.accounts.model.rest.BalanceSheet;
 import uk.gov.companieshouse.api.accounts.model.rest.CurrentPeriod;
 import uk.gov.companieshouse.api.accounts.model.rest.FixedAssets;
 import uk.gov.companieshouse.api.accounts.model.rest.PreviousPeriod;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.tangible.Cost;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.tangible.Depreciation;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.tangible.TangibleAssets;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.tangible.TangibleAssetsResource;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.tangible.Cost;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.tangible.Depreciation;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.tangible.TangibleAssets;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.tangible.TangibleAssetsResource;
 import uk.gov.companieshouse.api.accounts.model.validation.Errors;
 import uk.gov.companieshouse.api.accounts.service.CompanyService;
 import uk.gov.companieshouse.api.accounts.service.impl.CurrentPeriodService;
 import uk.gov.companieshouse.api.accounts.service.impl.PreviousPeriodService;
 import uk.gov.companieshouse.api.accounts.service.response.ResponseObject;
+import uk.gov.companieshouse.api.accounts.validation.AccountsResourceValidator;
+import uk.gov.companieshouse.api.accounts.validation.BaseValidator;
 import uk.gov.companieshouse.api.model.transaction.Transaction;
 
 @Component
-public class TangibleAssetsValidator extends BaseValidator implements CrossValidator<TangibleAssets> {
+public class TangibleAssetsValidator extends BaseValidator implements AccountsResourceValidator<TangibleAssets> {
 
     private CompanyService companyService;
 
@@ -63,7 +66,8 @@ public class TangibleAssetsValidator extends BaseValidator implements CrossValid
     private static final String NET_BOOK_VALUE_CURRENT_PERIOD = ".net_book_value_at_end_of_current_period";
     private static final String NET_BOOK_VALUE_PREVIOUS_PERIOD = ".net_book_value_at_end_of_previous_period";
 
-    public Errors validateTangibleAssets(TangibleAssets tangibleAssets, Transaction transaction, String companyAccountsId, HttpServletRequest request)
+    @Override
+    public Errors validateSubmission(TangibleAssets tangibleAssets, Transaction transaction, String companyAccountsId, HttpServletRequest request)
             throws DataException {
 
         Errors errors = new Errors();
@@ -943,8 +947,7 @@ public class TangibleAssetsValidator extends BaseValidator implements CrossValid
         return TANGIBLE_NOTE + "." + subResource.getJsonPath() + pathSuffix;
     }
 
-    @Override
-    public Errors crossValidate(TangibleAssets tangibleAssets,
+    private void crossValidate(TangibleAssets tangibleAssets,
                                 HttpServletRequest request,
                                 String companyAccountsId,
                                 Errors errors) throws DataException {
@@ -961,8 +964,6 @@ public class TangibleAssetsValidator extends BaseValidator implements CrossValid
         if (previousPeriodBalanceSheet != null) {
             crossValidatePreviousPeriod(errors, request, companyAccountsId, tangibleAssets);
         }
-
-        return errors;
     }
 
     private void crossValidateCurrentPeriod(Errors errors, HttpServletRequest request, String companyAccountsId,
@@ -1066,4 +1067,7 @@ public class TangibleAssetsValidator extends BaseValidator implements CrossValid
             return jsonPath;
         }
     }
+
+    @Override
+    public AccountsResource getAccountsResource() { return AccountsResource.SMALL_FULL_TANGIBLE_ASSETS; }
 }
